Move error class to init file
